# backend/app/routers/bookings.py
from fastapi import APIRouter
from pydantic import BaseModel
from datetime import datetime

router = APIRouter(prefix="/bookings", tags=["Bookings"])

class TrainBookingCreate(BaseModel):
    train_id: int
    departure_station_id: int
    arrival_station_id: int
    departure_time: datetime
    seat_class: str

class TrainBookingResponse(BaseModel):
    booking_id: str
    status: str

@router.post("/trains")
async def create_train_booking(booking: TrainBookingCreate):
    # Implement booking logic with Sirena-Travel API
    return {
        "booking_id": "RZD-2024-12345",
        "status": "подтверждено"
    }

class HotelBookingCreate(BaseModel):
    hotel_id: int
    room_type_id: int
    check_in_datetime: datetime
    check_out_datetime: datetime

class HotelBookingResponse(BaseModel):
    booking_id: str
    status: str

@router.post("/hotels")
async def create_hotel_booking(booking: HotelBookingCreate):
    # Implement booking logic with Ostrovok.ru API
    return {
        "booking_id": "OST-2024-67890",
        "status": "подтверждено"
    }